+++ @still/helper/form.js
@@ -6,37 +6,36 @@ class InParams {
     datasets = {}; 
     type; 
     placeholder; 
-    min; max; r
-    equired; 
+    min; max; required; warn;
     /** @type { 'number'|'alhpanumeric'|'text'|'email'|'phone'|'date'|'dateUS' } */
     validator;
 } ;
 
 export const FormHelper = {
-    newField(cmp, formRef, fieldName, value = null){
+    newField(cmp, formRef, fName, value = null){
         //Components is available globally from import { Components } from "../setup/components";        
-        Components.ref(cmp.cmpInternalId).setDynamicField(fieldName, value);
-        Components.obj().parseGetsAndSets(cmp, false,fieldName);
+        Components.ref(cmp.cmpInternalId).setDynamicField(fName, value);
+        Components.obj().parseGetsAndSets(cmp, false,fName);
         return {
             /** @param { InParams } params  */
-            getInput(params = inParams){
-                const {className, id, datasets = {}, type, placeholder, min, max, required, validator} = params;
+            input(params = inParams){
+                const {className, id, datasets = {}, type, placeholder, min, max, required, validator, warn} = params;
                 const datafields = Object.entries(datasets).map(([f,v]) => (`data-${f}="${v}"`)).join(' ');
                 const ftype=`type="${type || 'text'}"`, isOptList = ['radio','checkbox'].includes(type);
                 const hint = `${placeholder ? `placeholder="${placeholder}"` : ''}`;
                 const val = `${value ? `value="${value}"` : ''}`, _id = `${id ? `id="${id}"` : ''}`;
                 const mn = `${min ? `min="${min}"` : ''}`, mx = `${max ? `max="${max}"` : ''}`;
-                const req = `${required ? ' (required)="true" ' : ''}`;
-                const validatorClass = BehaviorComponent.setOnValueInput(req, cmp, fieldName, (formRef.name || null));
+                const req = `${required ? ' (required)="true" ' : ''}`, wrn = `${warn ? ` (validator-warn)="${warn}"` : ''}`;
+                const validatorClass = required ? BehaviorComponent.setOnValueInput(req, cmp, fName, (formRef.name || null)) : '';
                 const validateEvt = required ?
-                 `onkeyup="$still.component.ref('${cmp.cmpInternalId}').onValueInput(event,'${fieldName}',this, '${formRef.name}')"`
+                 `onkeyup="$still.component.ref('${cmp.cmpInternalId}').onValueInput(event,'${fName}',this, '${formRef.name}')"`
                  : '';
                 const vlidtor = `${validator ? `(validator)=${validator}`: ''}`;
                 const cmpId = this.cmpInternalId?.replace('/','').replace('@','');
                 const input = `
                     <input ${datafields}
-                        class="${genInputsClasses(validatorClass, cmpId, fieldName, val, isOptList)} ${cmp.cmpInternalId}-${fieldName} ${className || ''}"
-                        ${ftype} ${val} ${_id} ${req.trim()} ${hint} ${mn} ${mx}
+                        class="${genInputsClasses(validatorClass, cmpId, fName, val, isOptList)} ${cmp.cmpInternalId}-${fName} ${className || ''}"
+                        ${ftype} ${val} ${_id} ${req.trim()} ${wrn} ${hint} ${mn} ${mx}
                         ${validateEvt} ${vlidtor}
                     >
                 `;
@@ -44,7 +43,7 @@ export const FormHelper = {
                     add(cb = function(input){}, subContainer = null){
                         let cnt = cb(input), ctr = document.getElementById(formRef.formId);
                         if(subContainer) ctr = ctr.querySelector(subContainer);
-                        ctr.insertAdjacentHTML('beforeend', cnt || input);
+                        ctr.insertAdjacentHTML('beforeend', `<span>${cnt || input}</span>`);
                     },
                     element: input 
                 }